Feature/prepare for nutrientplans

+++ IIS-Projekat/IIS-Projekat/Models/FoodShare.cs
@@ -6,7 +6,7 @@ namespace IIS_Projekat.Models
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public long Id { get; set; }
-        public int Share { get; set; }
+        public float Share { get; set; }
         public Food Food { get; set; }
         public Recipe Recipe { get; set; }
         public DateTime CreatedDate { get; set; }
